[KP-22-23]Changed Module naming and sorted model into files
There are now Movies, Film, Countries, Genres module files
