+++ KinoPoisk/Core/Models/Countries.swift
@@ -1,0 +1,12 @@
+//
+//  Countries.swift
+//  KinoPoisk
+//
+//  Created by Philipp Zeppelin on 10.06.2023.
+//
+
+import Foundation
+
+struct Countries: Codable {
+    var country: String
+}
