+++ KinoPoisk/Core/Models/Genres.swift
@@ -1,0 +1,12 @@
+//
+//  Genres.swift
+//  KinoPoisk
+//
+//  Created by Philipp Zeppelin on 10.06.2023.
+//
+
+import Foundation
+
+struct Genres: Codable {
+    var genre: String
+}
