+++ KinoPoisk/Modules/DetailView/Presenter/DetailTableViewPresenter.swift
